[ADD] Select Manittee View 생성(#434)
