+++ Manito/Manito/Screens/Interaction/SelectManittoViewController.swift
@@ -7,43 +7,16 @@
 
 import UIKit
 
-import Gifu
-
 final class SelectManittoViewController: BaseViewController {
 
     private enum SelectionStage {
         case showJoystick, showCapsule, openName, openButton
     }
 
-    // MARK: - ui component
-
-    private let informationLabel: UILabel = {
-        let label = UILabel()
-        label.font = .font(.regular, ofSize: 20)
-        label.numberOfLines = 2
-        label.text = TextLiteral.selectManittoViewControllerInformationText
-        label.addLabelSpacing()
-        label.textAlignment = .center
-        return label
-    }()
-    private let nameLabel: UILabel = {
-        let label = UILabel()
-        label.font = .font(.regular, ofSize: 30)
-        return label
-    }()
-    private let confirmButton: MainButton = {
-        let button = MainButton()
-        button.title = TextLiteral.confirm
-        return button
-    }()
-    private let joystickBackgroundView: UIView = UIView()
-    private let joystickImageView: GIFImageView = GIFImageView(image: UIImage(named: ImageLiterals.gifJoystick))
-    private let openCapsuleImageView: GIFImageView = GIFImageView(image: UIImage(named: ImageLiterals.gifCapsule))
-
     // MARK: - property
 
-    var roomId: String?
-    var manitteeName: String?
+    private let roomId: String
+    var manitteeNickname: String?
     private var stageType: SelectionStage = .showJoystick {
         didSet {
             self.hiddenImageView()
@@ -51,89 +24,30 @@ final class SelectManittoViewController: BaseViewController {
         }
     }
 
+    // MARK: - init
+
+    init(roomId: String, manitteeNickname: String) {
+        self.roomId = roomId
+        self.manitteeNickname = manitteeNickname
+        super.init()
+    }
+
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
     // MARK: - life cycle
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.setupButtonAction()
-        self.setupSwipeGesture()
         self.setupGifImage()
         self.hiddenImageView()
     }
-
-    // MARK: - override
-
-    override func setupLayout() {
-        self.view.addSubview(self.joystickBackgroundView)
-        self.joystickBackgroundView.snp.makeConstraints {
-            $0.edges.equalTo(self.view.safeAreaLayoutGuide)
-        }
-
-        self.joystickBackgroundView.addSubview(self.joystickImageView)
-        self.joystickImageView.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.centerY.equalToSuperview().offset(-30)
-            $0.width.height.equalTo(140)
-        }
-
-        self.joystickBackgroundView.addSubview(self.informationLabel)
-        self.informationLabel.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.top.equalTo(self.joystickImageView.snp.bottom).offset(63)
-        }
-
-        self.view.addSubview(self.openCapsuleImageView)
-        self.openCapsuleImageView.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.centerY.equalToSuperview().offset(-30)
-            $0.width.equalTo(199)
-            $0.height.equalTo(285)
-        }
-
-        self.view.addSubview(self.nameLabel)
-        self.nameLabel.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.centerY.equalTo(self.openCapsuleImageView.snp.centerY)
-        }
-
-        self.view.addSubview(self.confirmButton)
-        self.confirmButton.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.bottom.equalTo(self.view.safeAreaLayoutGuide).inset(31)
-        }
-    }
-
-    override func configureUI() {
-        super.configureUI()
-
-        if let manittee = self.manitteeName {
-            self.nameLabel.text = manittee
-        }
-    }
+    
 
     // MARK: - func
 
-    private func setupButtonAction() {
-        let okAction = UIAction { [weak self] _ in
-            guard let presentingViewController = self?.presentingViewController as? UINavigationController,
-                  let roomId = self?.roomId
-            else { return }
-            let viewController = DetailingViewController(roomId: roomId)
-            presentingViewController.popViewController(animated: true)
-            presentingViewController.pushViewController(viewController, animated: false)
-            self?.dismiss(animated: true)
-        }
-        self.confirmButton.addAction(okAction, for: .touchUpInside)
-    }
-
-    private func setupSwipeGesture() {
-        let swipeLeft = UISwipeGestureRecognizer(target: self, action: #selector(self.respondToSwipeGesture(_:)))
-        let swipeRight = UISwipeGestureRecognizer(target: self, action: #selector(self.respondToSwipeGesture(_:)))
-        swipeLeft.direction = UISwipeGestureRecognizer.Direction.left
-        swipeRight.direction = UISwipeGestureRecognizer.Direction.right
-        self.joystickBackgroundView.addGestureRecognizer(swipeLeft)
-        self.joystickBackgroundView.addGestureRecognizer(swipeRight)
-    }
 
     private func setupGifImage() {
         switch stageType {
@@ -174,17 +88,4 @@ final class SelectManittoViewController: BaseViewController {
         }
     }
 
-    // MARK: - selector
-
-    @objc
-    private func respondToSwipeGesture(_ gesture: UIGestureRecognizer) {
-        if let swipeGesture = gesture as? UISwipeGestureRecognizer {
-            switch swipeGesture.direction {
-            case .left, .right:
-                self.stageType = .showCapsule
-            default:
-                break
-            }
-        }
-    }
 }
